fix warnings

+++ lib/http_perform.c
@@ -424,7 +424,7 @@ sf_bool STDCALL http_perform(CURL *curl,
         /* Check for errors */
         if (res != CURLE_OK) {
           if (res == CURLE_COULDNT_CONNECT && curl_retry_ctx.retry_count <
-                                              retry_on_curle_couldnt_connect_count)
+                                              (unsigned)retry_on_curle_couldnt_connect_count)
             {
               retry = SF_BOOLEAN_TRUE;
               uint32 next_sleep_in_secs = retry_ctx_next_sleep(&curl_retry_ctx);
@@ -437,8 +437,8 @@ sf_bool STDCALL http_perform(CURL *curl,
             } else if ((res == CURLE_OPERATION_TIMEDOUT) && (renew_timeout > 0)) {
               retry = SF_BOOLEAN_TRUE;
             } else if (res == CURLE_PARTIAL_FILE) {
-              if (((time(NULL) - elapsedRetryTime) < curl_retry_ctx.retry_timeout) &&
-                  ((retry_max_count <= 0) || (curl_retry_ctx.retry_count < retry_max_count)))
+              if (((uint64)(time(NULL) - elapsedRetryTime) < curl_retry_ctx.retry_timeout) &&
+                  ((retry_max_count <= 0) || (curl_retry_ctx.retry_count < (unsigned)retry_max_count)))
               {
                   uint32 next_sleep_in_secs = retry_ctx_next_sleep(&curl_retry_ctx);
                   log_debug(
@@ -497,8 +497,8 @@ sf_bool STDCALL http_perform(CURL *curl,
                 curl_retry_ctx.retry_reason = (uint32)http_code;
               }
               if (retry &&
-                  ((time(NULL) - elapsedRetryTime) < curl_retry_ctx.retry_timeout) &&
-                  ((retry_max_count <= 0) || (curl_retry_ctx.retry_count < retry_max_count)))
+                  ((uint64)(time(NULL) - elapsedRetryTime) < curl_retry_ctx.retry_timeout) &&
+                  ((retry_max_count <= 0) || (curl_retry_ctx.retry_count < (unsigned)retry_max_count)))
               {
                 uint32 next_sleep_in_secs = retry_ctx_next_sleep(&curl_retry_ctx);
                 log_debug(
